move cpa to its own file
